@@ -102,7 +102,7 @@ class HtmlWidget extends StatefulWidget {
     this.customWidgetBuilder,
     this.enableCaching,
     this.factoryBuilder,
-    Key? key,
+    super.key,
     this.onErrorBuilder,
     this.onLoadingBuilder,
     this.onTapImage,
@@ -110,8 +110,7 @@ class HtmlWidget extends StatefulWidget {
     List<dynamic>? rebuildTriggers,
     this.renderMode = RenderMode.column,
     this.textStyle,
-  })  : _rebuildTriggers = rebuildTriggers,
-        super(key: key);
+  }) : _rebuildTriggers = rebuildTriggers;
 
   @override
   State<HtmlWidget> createState() => HtmlWidgetState();
