@@ -837,7 +837,7 @@ Future<void> main() async {
 class _Golden extends StatelessWidget {
   final String html;
 
-  const _Golden(this.html, {Key? key}) : super(key: key);
+  const _Golden(this.html);
 
   @override
   Widget build(BuildContext _) => Scaffold(
