@@ -145,14 +145,20 @@ Future<void> main() async {
       const html =
           '<table border="1"><tbody><tr><td>Foo</td></tr></tbody></table>';
       final explained = await explain(tester, html, useExplainer: false);
-      expect(explained, contains('HtmlTable(borderSpacing: 2.0)'));
+      expect(
+        explained,
+        contains('HtmlTable(border: all(BorderSide), borderSpacing: 2.0)'),
+      );
     });
 
     testWidgets('renders style', (WidgetTester tester) async {
       const html = '<table style="border: 1px solid black"><tbody>'
           '<tr><td>Foo</td></tr></tbody></table>';
       final explained = await explain(tester, html, useExplainer: false);
-      expect(explained, contains('HtmlTable(borderSpacing: 2.0)'));
+      expect(
+        explained,
+        contains('HtmlTable(border: all(BorderSide), borderSpacing: 2.0)'),
+      );
     });
   });
 
@@ -225,8 +231,13 @@ Future<void> main() async {
       const html = '<table border="1" style="border-collapse: collapse"><tbody>'
           '<tr><td>Foo</td></tr>'
           '</tbody></table>';
-      final e = await explain(tester, html, useExplainer: false);
-      expect(e, contains('(borderCollapse: true, borderSpacing: 2.0)'));
+      final explained = await explain(tester, html, useExplainer: false);
+      expect(
+        explained,
+        contains(
+          'HtmlTable(border: all(BorderSide), borderCollapse: true, borderSpacing: 2.0)',
+        ),
+      );
     });
   });
 
@@ -1195,7 +1206,7 @@ Future<void> main() async {
 class _Golden extends StatelessWidget {
   final String contents;
 
-  const _Golden(this.contents, {Key? key}) : super(key: key);
+  const _Golden(this.contents);
 
   @override
   Widget build(BuildContext _) => Scaffold(
